requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.17.0
alembic==1.13.3
altair==5.4.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
exceptiongroup==1.2.2
filelock==3.16.1
flake8==5.0.4
flake8-isort==6.1.1
fonttools==4.54.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
magic-filter==1.0.12
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mccabe==0.7.0
mdurl==0.1.2
multidict==6.1.0
narwhals==1.13.2
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.0
protobuf==5.28.3
pyarrow==18.0.0
pycodestyle==2.9.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pyflakes==2.5.0
Pygments==2.18.0
pyparsing==3.2.0
pyproject-api==1.8.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-flake8==1.2.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tomli==2.0.2
tornado==6.4.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
watchdog==5.0.3
yarl==1.15.3
zipp==3.20.2
